[NB] Use uint16_t for payload length

+++ nadamq/src/PacketWriter.h
@@ -59,7 +59,7 @@ inline void write_packet(Stream &output, Packet const &packet) {
   uint8_t type_ = static_cast<uint8_t>(to_send.type());
   serialize_any(output, type_);
   if (to_send.type() == Packet::packet_type::DATA) {
-    serialize_any(output, static_cast<uint8_t>(to_send.payload_length_));
+    serialize_any(output, static_cast<uint16_t>(to_send.payload_length_));
     if (to_send.payload_length_ > 0) {
       output.write((stream_byte_type*)to_send.payload_buffer_,
                    to_send.payload_length_);
